Since the last commit I managed to implement a lot of features. I also forgor to commit yesterday so there will be a large commit today.
The create and edit horse buttons work fully now and correctly store the details of the horse in the horseDetails.txt
file. The details being name, symbol, confidence, colour, breed and accessories- when the horse is created
the average speed is defaulted to 15 and the win ratio is 0.5. This is just due to the method in which the
they are calculated.

I also add functionality to the view statistics button in the menu- this displays the statistics of every
horse thats been  created this includes their name, symbol, confidence, colour, breed, accessories, average
speed and win ratio.

One of the larger parts I implemented was the whole racing part of the project. This includes the window for creating
the race when create race is clicked on the menu bar; the slider for choosing the race length; the add horse button
which creates a new window with clickable labels of all the horses, which when clicked will add those horses into
the race; there is also a text area in the middle of the window which displays the horses that have already been added;
Finally there is the start race button that when clicked opens a new window with the animated race.

When the race is finished at either: win, draw or all lose, the indivisual horses statistics are updated like their
win ratio, average speed and confidence. The race details are also formated and stored in the raceDetails.txt file.
These details include the date and time the race happened, the status of the race: if it was a win, draw or all lose; the
length of the race, the participating horses, if they fell and the distance they moved.

I also added the view race history button in the race sub menu which displays the details, that I covered in the
last paragraph, of every past race by reading it from the raceDetails.txt file in a new window.

The only thing I have left to do is the betting part of the project, fixing a problem with the race window not being
able to close properly; the code formatting as it is a bit messy; and the aesthetics of the project as at the moment
it is very basic.