So I managed to get the race window properly closing, so now its independent if the create horses window and can be moved around and closed whenever. I also had to fix some bugs relating to the updating of the horses statistics after the race, I found out that when I created each horse the average speed was being set to 15, and not 15.0. This led the program to think that they were not the same line and didnt update them.
Other than those I changes I've removed some of the print statements that I was using to debug the program and added
a scroll bar to the add horse window as if there were too many it wouldn't show bottom horses.

Next plans are to make the GUI look a lot more appealing and do some github related stuff.